Bitti.

+++ src/cofr_ornek2/OdemeYap.java
@@ -1,10 +1,10 @@
 package cofr_ornek2;
 
 public class OdemeYap {
-	public float ucret;
+	public double ucret;
 	public Yontemler yontemSec;
 	
-	public OdemeYap(float ucret, Yontemler yontem) {
+	public OdemeYap(double ucret, Yontemler yontem) {
 		this.ucret = ucret;
 		this.yontemSec = yontem;
 	}
